refactor(topology): fold the gossip task into the behaviour as a sync engine
The gossip task was a synchronous state machine wrapped in a select loop:
not one await in its logic, and both of its channels were
behaviour-to-behaviour round trips (every input send site a behaviour
method, the only action consumer the behaviour's own poll). The detour
produced the wake gap, silent drop-on-full in both directions, and a
depth shadow synchronized by droppable events.

Fold it into a GossipEngine owned by the behaviour: the five input
variants become methods returning the actions to broadcast, applied in
the same call through the existing broadcast machinery, so gossip can
neither lag nor drop by construction. The refresh interval is polled in
poll beside the dial interval; the health-check sleeps become a deadline
queue armed by one boxed timer, which deletes the cancellation set and
its cleanup heuristic and removes the MaybeSend future fork from gossip.
The depth shadow is replaced by passing the published depth at each call
site. The input enum, both channels, the handle, the spawn and the
buffered pre-spawn semantics are deleted; the action struct survives as
the engine's tested return vocabulary. The lifecycle broadcast bridge
from the libp2p-free peer manager is untouched.

The crate rustdoc now states the task-placement rule: detached tasks are
for heavy or coalescible compute and real I/O; reactive policy and its
timers live in the behaviour's poll; cross-boundary event flows are
streams polled with a registered waker.
